Workaround cspell bug when cwd contains spaces
... or probably any chars that gets percent encoded when transformed
to a URI.

cspell translates the given path to a URI but the reporter has a bug
where it compares the normal working dir with the percent encoded one.
These two never match obviously and then cspell prints the complete
percent encoded URI as the filename.

We thus (a) teach the regex about these stdin filenames, (b) teach
`is_stdin_filename` to always return `True` because the filename
the reporter reports is not a filename that can found on the filesystem
but an URI.  If we return true, SublimeLinter core assumes the current
view